Adjust test cases to support automated test
