+++ .github/workflows/go.yml
@@ -4,7 +4,7 @@
 name: Go
 
 on:
-  [push]
+  [push, pull_request, workflow_dispatch]
 
 jobs:
