+++ tests/udp.go
@@ -5,46 +5,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"net"
-	"strings"
-	"testing"
 	"time"
 
 	"github.com/txthinking/socks5"
 )
-
-// go test -v -run=TestUDPByProxy
-func TestUDPByProxy(t *testing.T) {
-	tunaNode, err := getTunaNode()
-	if err != nil {
-		fmt.Printf("getTunaNode err %v\n", err)
-		return
-	}
-
-	tuna, udp, tun := true, true, false
-	go func() {
-		err := startNconnect("server.json", tuna, udp, tun, tunaNode)
-		if err != nil {
-			fmt.Printf("start nconnect server err: %v\n", err)
-			return
-		}
-	}()
-
-	time.Sleep(15 * time.Second)
-
-	go func() {
-		err := startNconnect("client.json", tuna, udp, tun, nil)
-		if err != nil {
-			fmt.Printf("start nconnect client err: %v\n", err)
-			return
-		}
-	}()
-
-	time.Sleep(15 * time.Second)
-
-	go StartUDPClient()
-
-	waitFor(ch, exited)
-}
 
 func StartUdpServer() error {
 	a, err := net.ResolveUDPAddr("udp", udpServerAddr)
@@ -65,7 +29,6 @@ func StartUdpServer() error {
 			fmt.Printf("StartUdpServer.ReadFromUDP err: %v\n", err)
 			break
 		}
-		fmt.Printf("UDP Server got: %v\n", string(b[:n]))
 
 		time.Sleep(100 * time.Millisecond)
 		_, _, err = udpServer.WriteMsgUDP(b[:n], nil, addr)
@@ -73,26 +36,19 @@ func StartUdpServer() error {
 			fmt.Printf("StartUdpServer.WriteMsgUDP err: %v\n", err)
 			break
 		}
-
-		if strings.Contains(string(b[:n]), udpClientExited) {
-			break
-		}
 	}
 
-	ch <- udpServerExited
 	return nil
 }
 
 func StartUDPClient() error {
 	s5c, err := socks5.NewClient(proxyAddr, "", "", 0, 60)
 	if err != nil {
-		ch <- udpClientExited
 		return err
 	}
 	uc, err := s5c.Dial("udp", udpServerAddr)
 	if err != nil {
 		fmt.Println("StartUDPClient.s5c.Dial err: ", err)
-		ch <- udpClientExited
 		return err
 	}
 	defer uc.Close()
@@ -117,11 +73,6 @@ func StartUDPClient() error {
 			break
 		}
 	}
-
-	uc.Write([]byte(udpClientExited))
-	time.Sleep(time.Second)
-
-	ch <- udpClientExited
 
 	return nil
 }
