+++ tests/web.go
@@ -8,43 +8,13 @@ import (
 	"log"
 	"net/http"
 	"net/url"
-	"testing"
 	"time"
 )
-
-// go test -v -run=TestHttpByProxy
-func TestHttpByProxy(t *testing.T) {
-	tuna, udp, tun := true, true, false
-
-	go func() {
-		err := startNconnect("server.json", tuna, udp, tun, nil)
-		if err != nil {
-			fmt.Printf("start nconnect server err: %v\n", err)
-			return
-		}
-	}()
-
-	time.Sleep(15 * time.Second)
-
-	go func() {
-		err := startNconnect("client.json", tuna, udp, tun, nil)
-		if err != nil {
-			fmt.Printf("start nconnect client err: %v\n", err)
-			return
-		}
-	}()
-	time.Sleep(15 * time.Second)
-
-	go StartWebClient()
-
-	waitFor(ch, exited)
-}
 
 func StartWebServer() error {
 	http.HandleFunc("/httpEcho", httpEcho)
 	fmt.Println("WEB server is serving at ", httpServerAddr)
 	if err := http.ListenAndServe(httpServerAddr, nil); err != nil {
-		ch <- webServerExited
 		log.Fatal(err)
 	}
 
@@ -64,7 +34,6 @@ func httpEcho(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	fmt.Println("web server got user:", user)
 	b, _ := json.Marshal(user)
 	w.Write(b)
 }
@@ -122,7 +91,6 @@ func StartWebClient() error {
 			break
 		}
 
-		fmt.Printf("respUser %+v\n", respUser)
 		if respUser.Age != user.Age {
 			fmt.Printf("StartWebClient got wrong response, sent %+v, recv %+v\n", user, respUser)
 			break
@@ -130,7 +98,58 @@ func StartWebClient() error {
 	}
 
 	time.Sleep(time.Second)
-	ch <- webClientExited
+
+	return nil
+}
+
+func StartTunWebClient() error {
+	httpClient := &http.Client{
+		Timeout: 10 * time.Second,
+	}
+
+	user := &Person{Name: "http_tun_boy", Age: 0}
+	b := new(bytes.Buffer)
+
+	for i := 0; i < 10; i++ {
+		user.Age++
+		err := json.NewEncoder(b).Encode(user)
+		if err != nil {
+			fmt.Printf("StartWebClient.Encode err: %v\n", err)
+			break
+		}
+		req, err := http.NewRequest(http.MethodPost, httpServiceUrl, b)
+		req.Header.Set("Content-type", "application/json")
+
+		if err != nil {
+			fmt.Printf("StartWebClient.http.NewRequest err: %v\n", err)
+			break
+		}
+
+		resp, err := httpClient.Do(req)
+		if err != nil {
+			fmt.Printf("StartWebClient.http.Do err: %v\n", err)
+			break
+		}
+		defer resp.Body.Close()
+
+		body, err := io.ReadAll(resp.Body)
+		if err != nil {
+			fmt.Printf("StartWebClient.io.ReadAll err: %v\n", err)
+			break
+		}
+
+		respUser := &Person{}
+		err = json.Unmarshal(body, respUser)
+		if err != nil {
+			fmt.Printf("StartWebClient.json.Unmarshal err: %v\n", err)
+			break
+		}
+
+		if respUser.Age != user.Age {
+			fmt.Printf("StartWebClient got wrong response, sent %+v, recv %+v\n", user, respUser)
+			break
+		}
+	}
 
 	return nil
 }
